requirements.txt
pandas
xlrd
python-slugify
regex